+++ examples/autonat/src/bin/autonat_server.rs
@@ -24,7 +24,7 @@ use clap::Parser;
 use futures::StreamExt;
 use libp2p::core::{multiaddr::Protocol, Multiaddr};
 use libp2p::swarm::{NetworkBehaviour, SwarmEvent};
-use libp2p::{autonat, identify, identity};
+use libp2p::{autonat, identify, identity, noise, tcp, yamux};
 use std::error::Error;
 use std::net::Ipv4Addr;
 
@@ -43,8 +43,11 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     let mut swarm = libp2p::SwarmBuilder::with_new_identity()
         .with_async_std()
-        .with_tcp()
-        .with_noise()?
+        .with_tcp(
+            tcp::Config::default(),
+            noise::Config::new,
+            yamux::Config::default,
+        )?
         .with_behaviour(|key| Behaviour::new(key.public()))?
         .build();